Pubsub playback

+++ misc/gcloud_pubsub_mirror/mirror.py
@@ -17,6 +17,7 @@ from google.cloud import pubsub_v1
 
 messages_processed = 0
 messages = False
+publish_futures = []
 
 def is_file_project(target: str) -> bool:
   return target == '//'
@@ -29,8 +30,21 @@ def subscribe_callback(message: pubsub_v1.subscriber.message.Message) -> None:
     print(f'{messages_processed} messages processed')
   message.ack()
 
+def get_publish_callback(publish_future: pubsub_v1.publisher.futures.Future):
+  global publish_futures
+  def callback(publish_future: pubsub_v1.publisher.futures.Future) -> None:
+    try:
+      # Wait 60 seconds for the publish call to succeed.
+      publish_future.result(timeout=60)
+    except futures.TimeoutError:
+      print(f"Publish timed out.")
+  return callback
+
 def topic_publisher(publisher, topic_path, message):
-  publisher.publish(topic_path, message.data, **message.attributes)
+  print('Publishing', topic_path, message.publish_time)
+  publish_future = publisher.publish(topic_path, message.data, **message.attributes)
+  publish_future.add_done_callback(get_publish_callback(publish_future))
+  publish_futures.append(publish_future)
 
 def file_publisher(path: str, message):
   fullstamp = message.publish_time.isoformat() + 'Z'
@@ -47,12 +61,15 @@ def file_publisher(path: str, message):
     "attributes": dict(message.attributes)
   }
 
-  print('Writing ' + file_name)
-  with open(file_name, "w") as outfile:
-    outfile.write(json.dumps(message_dict, indent=2))
+  if not os.path.exists(file_name):
+    print('Writing ' + file_name)
+    with open(file_name, "w") as outfile:
+      outfile.write(json.dumps(message_dict, indent=2))
+  else:
+    print('Skipping ' + file_name)
 
 def load_messages(path: str):
-  print('loading ' + path)
+  print('Processing ' + path)
   if os.path.isfile(path):
     with open(path, 'r') as json_file:
       yield json.load(json_file)
@@ -136,3 +153,6 @@ while has_messages:
 if future:
   future.cancel()
   future.result()
+
+print('Waiting for message publishing to complete...')
+futures.wait(publish_futures, return_when=futures.ALL_COMPLETED)
